requirements.txt
# ── InsightOps Intelligence — Dependencies ─────────────────

# CV Pipeline
ultralytics>=8.0
insightface>=0.7
deep-sort-realtime>=1.3
opencv-python>=4.8
onnxruntime>=1.16

# Data layer
pandas>=2.0
numpy>=1.24

# Azure
azure-storage-blob>=12.0

# AI narration
openai>=1.0
python-dotenv>=1.0

# Output
reportlab>=4.0
streamlit>=1.30
Pillow>=10.0
